adds controllers
